hack: add minio instance for testing s3 cache to compose
Adds a profile `remotecache` for a minio instance to be used as an s3
importer and exporter. Can be used by using the following:

	--cache-to type=s3,region=xyz,bucket=mybucket,name=buildkit,endpoint_url=http://minio:9000,access_key_id=abc,secret_access_key=def,use_path_style=true

Some of these parameters are just required by the s3 endpoint and aren't
significant for minio. The region doesn't matter. The access key and
secret key for the minio instance defaults to `minioadmin:minioadmin`
but can be changed. The bucket has to be created through the web UI. The
name is the name of the project or whatever name you want.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>